requirements.txt
glob2
keras
matplotlib
numpy
opencv-python
pandas
Pillow
pydot
scikit-learn
seaborn
tensorflow